requirements.txt
apache-airflow-providers-amazon==8.6.0
apache-airflow-providers-snowflake==5.0.0
boto3==1.28.17
snowflake-snowpark-python